requirements.txt
transformers>=4.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pandas>=2.0.0
requests>=2.30.0
